오류 수정
Entity랑 Repository 사이의 오류 수정

오류 발생 이유
Entity와 Repository 사이의 notiId 데이터 타입을 통일시키지 않음

Long으로 맞춤

+++ demo/src/main/java/com/example/Attendance/entity/Notification.java
@@ -16,7 +16,7 @@ public class Notification {
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "notification_seq")
 	@SequenceGenerator(name="notification_seq", sequenceName="NOTIFICATION_SEQ", allocationSize = 1)
-	private int notiId;
+	private Long notiId;
 	
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "board_id")
